move facilitator guide into facilitators folder
